Moved active zuora objects into their own directory to avoid confusion
