+++ lib/active_zuora.rb
@@ -1,28 +1,29 @@
 require 'delegate'
 require 'zuora_client'
-require 'zuora/zobject'
-require 'zuora/account'
-require 'zuora/bill_run'
-require 'zuora/contact'
-require 'zuora/invoice'
-require 'zuora/invoice_item_adjustment'
-require 'zuora/invoice_payment'
-require 'zuora/payment'
-require 'zuora/payment_method'
-require 'zuora/product'
-require 'zuora/product_rate_plan'
-require 'zuora/product_rate_plan_charge'
-require 'zuora/product_rate_plan_charge_tier'
-require 'zuora/product_rate_plan_charge_tier_data'
-require 'zuora/rate_plan'
-require 'zuora/rate_plan_charge'
-require 'zuora/rate_plan_charge_tier'
-require 'zuora/rate_plan_data'
-require 'zuora/refund'
-require 'zuora/subscribe_options'
-require 'zuora/subscribe_request'
-require 'zuora/subscribe_with_existing_account_request'
-require 'zuora/subscription'
-require 'zuora/subscription_data'
-require 'zuora/usage'
+require 'active_zuora/zobject'
+require 'active_zuora/account'
+require 'active_zuora/bill_run'
+require 'active_zuora/contact'
+require 'active_zuora/invoice'
+require 'active_zuora/invoice_item_adjustment'
+require 'active_zuora/invoice_payment'
+require 'active_zuora/payment'
+require 'active_zuora/payment_method'
+require 'active_zuora/product'
+require 'active_zuora/product_rate_plan'
+require 'active_zuora/product_rate_plan_charge'
+require 'active_zuora/product_rate_plan_charge_tier'
+require 'active_zuora/product_rate_plan_charge_tier_data'
+require 'active_zuora/rate_plan'
+require 'active_zuora/rate_plan_charge'
+require 'active_zuora/rate_plan_charge_tier'
+require 'active_zuora/rate_plan_data'
+require 'active_zuora/refund'
+require 'active_zuora/subscribe_options'
+require 'active_zuora/subscribe_request'
+require 'active_zuora/subscribe_with_existing_account_request'
+require 'active_zuora/subscription'
+require 'active_zuora/subscription_data'
+require 'active_zuora/usage'
+
