+++ lib/active_zuora/product.rb
