+++ lib/active_zuora/zobject.rb
